restructuring application to sustainable modern python package design
