+++ jericho/README.md
@@ -9,6 +9,3 @@ jericho is the central server for book metadata and review information
     - ORM: [SQLModel (SQLAlchemy + Pydantic)](https://sqlmodel.tiangolo.com/)
     - Migrations: [Alembic](https://alembic.sqlalchemy.org/en/latest/)
 - Relational database: PostgreSQL 15
-- Swagger API
-  - Codegen: [fastapi-code-generator](https://github.com/koxudaxi/fastapi-code-generator)
-  -
