+++ jericho/app/.gitignore
