+++ jericho/app/pyproject.toml
@@ -1,16 +1,16 @@
 [tool.poetry]
-name = "jericho"
+name = "app"
 version = "0.1.0"
 description = ""
 authors = ["Arden Chew <achew@grailbio.com>"]
 readme = "README.md"
-packages = [{ include = "**/*" }]
 
 [tool.poetry.dependencies]
 python = "^3.11"
-fastapi = "^0.101.0"
-uvicorn = {extras = ["standard"], version = "^0.23.2"}
+fastapi = "^0.101.1"
 sqlmodel = "^0.0.8"
+uvicorn = "^0.23.2"
+psycopg2 = "^2.9.7"
 
 
 [build-system]
