+++ jericho/app/src/routers/internal.py
@@ -1,0 +1,10 @@
+from fastapi import APIRouter
+
+router = APIRouter(
+    tags=["internal"],
+)
+
+
+@router.get("/health")
+async def health():
+    return {"message": "OK"}
