+++ jericho/env.bash
@@ -1,3 +1,3 @@
-alias start-server='cd $BIBLI/jericho/app;  poetry run uvicorn app.main:app --reload'
+alias start-server='cd $BIBLI/jericho/app;  poetry run uvicorn main:app --reload'
 
 alias pg-jericho-local='psql postgres -d jericho'
